Made body measurements return double instead of integer to accommodate foot sizes

+++ src/application/BodyPart.java
@@ -3,6 +3,6 @@ package application;
 import application.clothingModel.InvalidSizeException;
 
 public interface BodyPart {
-    public int getMeasurement(ClothingType clothingType) throws InvalidSizeException;
+    public double getMeasurement(ClothingType clothingType) throws InvalidSizeException;
     public void setMeasurement(Measurement measurement);
 }
